Added specs for the Cloudtasker ActiveJob Worker
